Enum stuff from Parin.

+++ benchmarks/array_append_remove/others/app_nim.nim
@@ -8,8 +8,8 @@ if n == 0:
     echo "Usage: array_append_remove <count>"
     quit(0)
 
-var array: seq[int]
+var array: seq[int32]
 for i in 0..<n:
-    array.add(i)
+    array.add(int32(i))
 while array.len != 0:
     discard array.pop()
